+++ src/sync/history/model.rs
@@ -26,3 +26,12 @@ pub struct HistoryList {
     #[serde(rename = "resetAll")]
     pub reset_all: Option<bool>,
 }
+
+impl crate::sync::common::Model for History {
+    fn get_id(&self) -> i32 {
+        self.id
+    }
+    fn get_updated_at(&self) -> i64 {
+        self.updated_at
+    }
+}