+++ src/sync/mod.rs
@@ -1,4 +1,6 @@
+pub mod common;
 pub mod history;
+pub mod live;
 pub mod manga;
 pub mod settings;
 pub mod update;